requirements.txt
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.41.0
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-storages==1.9.1
docutils==0.15.2
fabric==2.5.0
gunicorn==20.0.4
idna==2.9
invoke==1.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
Markdown==3.2.1
MarkupSafe==1.1.1
oauthlib==3.1.0
paramiko==2.7.1
Pillow==6.2.1
prettytable==0.7.2
psycopg2==2.8.4
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
urllib3==1.25.8

